[netcfg] Improve comments on default serde fields
Add comments to the custom structs in Netcfg with non-obvious default
behaviors to make it easier to understand there is custom default
behavior. This makes it easier to determine the default behavior when a
field is not specified in a netcfg configuration.

Fixed: 430307900
Change-Id: I5b095ddf3e279860cfad5bfb74d0ac25ecbc3b2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317366
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>