[vim3] Reduce saradc 2nd mmio range size
By mistake it took the entire length of the bus, which includes the
serial device that the kernel is claiming. A cursory look into the code
and bootloader devicetree, reveals that only one page is necessary for
this device.

This unblocks the kernel adding the uart range to the MMIO deny list.

Change-Id: Ib5b682ac34f57ab8fac4a4e65df4bc97a8d8ddf5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1316627
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>