[socket-proxy] Mark component as crtitical
When the socket-proxy component is stopped, the device should reboot.
This is required to ensure that sockets do not continue to be created
with the old closed socket provider channels. `fdio`, the primary
mechanism for using FIDL protocols over Zircon channels, does not make
any attempt to reconnect to the protocol after there are errors on the
channel. This means that even if the socket-proxy component restarts and
is functional again, all clients that use `fdio` for sockets will not be
able to open a socket. This leaves the device restarting the only
current method of recovery.

Bug: 417568023
Change-Id: I5314f4485ddd658aa693638b260fa363543fce78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276064
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>