[test][thread] Extend a timeout in a core thread test.
The timeouts being used in the core test which checks to see if the
ZX_THREAD_SUSPENDED signal works properly when used with ports were
quite short (only 100mSec) for running in a nested VM environment such
as CI/CQ.  A flake recently showed up which reported this timeout
being hit, but that does not seem particularly hard to believe given
the async nature of a thread suspend operation and the short timeout
here.

Change the timeout from 100 mSec to 180 seconds.  If we are unable to
suspend a thread in 180 seconds, we probably should investigate, even
though it is _technically_ possible for this to happen in a VM
environment.

Also, eliminate the test hook which tested that this functionality
worked properly for repeating async waits.  Repeating async waits have
not existed for a long time in the system, and the test was just doing
the exact same thing that the single-async-wait test was doing.  No
point in keeping around two copies.

Fixed: 431281599
Change-Id: Iaf789c64904b8be834d76c03866b9417be8b93fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1320724
Commit-Queue: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>