[kernel] s/ALIGN/ROUND in <align.h>
In this change we replace all usages of
* ALIGN with ROUNDUP
* PAGE_ALIGN with ROUNDUP_PAGE_SIZE
* IS_PAGE_ALIGNED with a new IS_PAGE_ROUNDED

Why? It was reasoned that it's more haphazard to have two aliases for
the same thing (ALIGN & ROUNDUP), especially when the connotative
distinction is arguable and subtle. Given that "round" is agreed as
being more general than "align" (which might be regarded as being for
addresses and powersof two), we settle on the ROUND spellings.

This change also has the benefit is clarifying that PAGE_ALIGN aligns
*up* in its replacement with ROUNDUP_PAGE_SIZE.

Change-Id: Ia3a804e2b3a05f3ce26081c1b42cfd76633c6294
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317744
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>