[kernel][vm] Trigger LRU when re-enabling aging
When running kernel unittests aging is sometimes disabled to simplify
the test implementation. These tests can additionally manipulate the
page queues and leave them in a state where the LRU needs processing,
but aging is disabled and so it is not triggered. When aging is
re-enabled the LRU thread will not get signalled until the MRU
actually needs processing, which depending on the system state and
configuration could be immediately or arbitrarily far into the future.
If it is far in the future then the LRU thread will emit a warning
message about it not being signalled despite the queue needing
processing.

Overall this issue is completely benign. When the MRU thread actually
needs space it will then issue the signal, and so the LRU thread will
eventually get signalled. Furthermore this only happens when running
kernel unittests that disable/enable aging. The warning message is
confusing though, so this change ensures that the page queues end up
in a more expected state after aging is re-enabled.

Change-Id: I7439ba01a004532282b2e062a8c6cf1b9db4dce2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1316764
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>