[starnix][functionfs] Support events
Theory: When the driver is torn down, adbd is not getting the hint
that it needs to reset state. Thus, when the connection comes back
up, adbd still thinks its in its old state and not sending data
that the server expects on start up.

This change signals adbd whenever the driver is torn down and
restarts adbd.

Bug: b/428551639

Test: tested `adb root/unroot`, USB plug/unplug. run through CQ.

Change-Id: I3c45aefd062d9d7cedacd64e9e9f14740b5906ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1320626
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>