[archivist][reader] Provide per-tree per-branch Inspect timeouts
Archivist was experiencing racy timeouts between the tree snapshot
as a whole, and individual slow lazy nodes. This change puts the
timeout on every individual fuchsia.inspect.Tree connection, so
that timeouts result in pruning behavior, rather than dropping
data that **can** be read.

This is achieved by managing the queue for reading the tree data in
a FuturesUnordered, such that the structure of the tree is discovered
independently from inflating the nodes.

This CL also fixes the Archivist integration test that should have
caught this regression in the first place.

The test previously passed because the lazy data was being written to a
different Inspector than all the other data. This was a quirk of the
puppet API. I modified the FIDL to have lazy nodes be part of the
InspectWriter protocol, so they can be added to an existing Inspector
rather than started as a new tree.

Fixed: 409822059
Change-Id: I587980383f47c0d73e02104ec2531a73c473f50d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1256184
Commit-Queue: Clayton McCray <claytonmccray@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Sam Balana <sbalana@google.com>
Fuchsia-Auto-Submit: Clayton McCray <claytonmccray@google.com>