[kernel][vm] Only move anonymous pages
Ensures that only pages actually in an anonymous queue get moved to an
anonymous zero fork queue. This indirectly solves the problem of the
previous MoveToAnonymousZeroFork invocation having different conditions
to this in the MoveToNotPinnedLocked method. Without this change the
page could be placed in the high priority queue, due to the VMO being
high priority, but then moved to the anonymous zero fork queue, which
is actually the reclaimable queue. This would cause high priority
pages, which are not reclaimable, to end up in the reclamation queues.

Bug: 423900604
Change-Id: I4c600e272742365043600cb4222bd80913ee0c52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312645
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>