[starnix] Count old pager reclaimable memory as available
In most platforms of interest, VMO discardable memory is tiny ; however,
a large pool of memory is used by pager memory, which is reclaimable by
Zircon. Old pager memory is easily reclaimable, while newer pager memory
is likely in use and reclaiming it would have a large impact on
performance.

To be more in line with the MemAvailable original purpose, this change
is adding the amount of old pager reclaimable memory to the amount of
memory reported as available, in addition to free and discardable
memory.

Change-Id: Ibdba9eaa4f29ca492ac6ead2ba90623193705d78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1321445
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Étienne J. Membrives <etiennej@google.com>