[starnix-syscalls-test] Deflake socket_test
The test became flaky because tearing down of the device is async and
the lingering caused the next test to receive unexpected netlink
messages that confuses the next test. This change fixes this by waiting
for the device to disappear in the tear down path.

The second problem is the assertion for not finding any routes should
be outside the while loop. Otherwise if the netstack responds to the
dump request in multiple messages, it can cause the test to flake.

Multiply: fuchsia-pkg://fuchsia.com/starnix_syscalls_cpp_tests#meta/socket_test.cm
Bug: 431142238
Change-Id: I6a6a6964b018ae9ea2f5c5bd7932ed3afbb44976
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1321602
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>