[tefmocheck] Make PSY checks fail if observed
There are currently no tests that can adequately catch when the daemon
does not start as the fallback paths report valid-looking results when
it is not present. Until we can shore up those paths, fail on this
observation as it's an indication that something is broken.

Bug: 431830276
Change-Id: Ib757f249e60b00a3556c9573918235a4e93f72ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1321788
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>
Fuchsia-Auto-Submit: Prashanth Swaminathan <prashanthsw@google.com>