[kernel][psci] Determine StateType of power_state values
Each CPU_SUSPEND power_state has a StateType that indicates whether it
is a powerdown state or a standby/retention state.  This commit
updates Zircon PSCI initialization to determine if the PSCI
implementation is using the original format or the extended ID format,
and then given the format, determines if a given power_state is a
powerdown state or not.

This commit also changes the behavior of platform_cpu_suspend to make
use of deeper-than-core power states when available.

Change spelling of noun "powerdown" to match that of the PSCI spec.

Update some bug numbers in TODO comments.

Sorrel PSCI boot spew now reads,

```
[00000.000] 00000:01025> PSCI version 1.1
[00000.000] 00000:01025> PSCI supported features:
[00000.000] 00000:01025>        CPU_SUSPEND (0x3)
[00000.000] 00000:01025>        CPU_OFF (0x0)
[00000.000] 00000:01025>        CPU_ON (0x0)
[00000.000] 00000:01025>        AFFINITY_INFO (0x0)
[00000.000] 00000:01025>        SYSTEM_OFF (0x0)
[00000.000] 00000:01025>        SYSTEM_RESET (0x0)
[00000.000] 00000:01025>        SYSTEM_RESET2 (0x0)
[00000.000] 00000:01025>        CPU_FREEZE (0x0)
[00000.000] 00000:01025>        SYSTEM_SUSPEND (0x0)
[00000.000] 00000:01025>        PSCI_SET_SUSPEND_MODE (0x0)
[00000.000] 00000:01025>        MEM_PROTECT (0x0)
[00000.000] 00000:01025>        MEM_PROTECT_RANGE (0x0)
[00000.000] 00000:01025>        PSCI64_SMCCC_VERSION (0x0)
[00000.000] 00000:01025> PSCI power_state format: extended
[00000.000] 00000:01025> PSCI power_states:
[00000.000] 00000:01025>   id=0x0, power_state=0x40000003, flags=0x1, powerdown=1
[00000.000] 00000:01025>     entry_latency_us=297, exit_latency_us=324, min_residency_us=1110
[00000.000] 00000:01025>   id=0x1, power_state=0x41000043, flags=0x2, powerdown=1
[00000.000] 00000:01025>     entry_latency_us=800, exit_latency_us=2118, min_residency_us=7376
```

Bug: 417558115
Change-Id: Ic6505d8c726594a9aa7fabaee7271d39169a060c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1313116
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>