[bazel] sdk/lib/utf-utils/BUILD.bazel
This file is only suitable to build for host, as it lacks
a Fuchsia-specific dependency in //zircon/system/ulib/zx
which has no BUILD.bazel file of its own for now.

Note that it is not possible to convert this to use
idk_cc_source_library() because this wants the top-level
//sdk source directory to be added to the include path
when compiling the library (or any of its dependents),
and this is not supported by `include_base` for now.

Bug: 426136195
Change-Id: I780a9e5e79e848cd39333d84471cbda9b39d1669
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1303048
Reviewed-by: David Dorwin <ddorwin@google.com>
Commit-Queue: David Turner <digit@google.com>