[kernel][interrupt] make interrupt handlers fit::inline_functions
It has always been that interrupt handlers at the kernel level have been
a function pointer plus a void * for context. This is pretty much
exactly what a fit::inline_function would allow for, so more or less
mechanically replace the api for one with the other, including all of
the driver code that is using it.

Change-Id: I66ad18fb01f7e4ecf481e484368c95110392cfa0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1302824
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>