[wlan][policy] Use next roam enabled time
This change adds a new field to StationaryMonitor,
`next_roaming_enabled_time`, which represents the earliest time that
roaming can be considered. When we have reached the maximum number of
roam attempts for a day, this gets set to the timestamp of the oldest
roam attempt + 24hrs.

This 1) allows  us to log when roaming has been disabled due to reaching
the maximum attempts threshold, and 2) prevents us from needing to query
for the recent roam events every time a signal report comes in.

Bug: None
Test: Added unit test

Change-Id: I9d6a28b3a824d7f56eb0ba5c37aa67e2072ea9d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317725
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>
Commit-Queue: Hayden Nix <haydennix@google.com>