[assembly] Wire up `zbi_extra_items`
From `board_configuration` in GN to assembly `zbi.rs` zbi tool wrapper.

This allows a board to provide custom ZBI items to included in the form
of a container. The template will merge all items through a ZBI tool
invocation through a GN action, and hand over this non-bootable ZBI to
assembly to be incoorporated into the final ZBI.

This is a multi-product build compatible way of achieving this.

Change-Id: I3a78ecd73b734f9807f45016bce2ec16b759c9cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1319749
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>