[sorrel][uart] Fix GENI TX IRQ handling.
Fix a handful of issues in the GENI serial driver which were causing
some suboptimal IRQ behavior.

Background:
This HW unit seems to be something of a "flexible serial unit".  It
has firmware loaded into it, and can handle standard UART transmission
in addition to things like I2C, SPI, and so on.

The UART TX interface is structured to transmit when it has received
an explicit command to do so.  Users are supposed to write the size
(in bytes) of a job into one register, then trigger transmission by
sending a GO command to the UART. The FIFO interface, however, can be
larger than a single byte  On Sorrel, the HW seems to expect data to
be delivered to it in 4-byte words.  This more or less necessitates
the command interface; if we have a non-multiple of 4 bytes to send,
how else does the HW (or micro-controller) know that the last word
written into the FIFO is a partial word?  The byte-length written
before issuing the command is what allows the HW to know, but
unfortunately adds some small amount of inefficiency as we need to
wait for a command to be fully completed before starting a new one.

So, there are two different interrupts we can use to drive UART TX.
The "low water mark" IRQ, and the "command complete" IRQ.  For any
given command that we start, we want to keep packing the FIFO as much
as we can, and then wait for the low water mark IRQ to fire before
packing more.  Once we have packed all of the bytes for the job into
the FIFO, we want to wait for the command complete IRQ before starting
the next command.

Issues and Fixes:
This biggest issue here had to do with the handling of these two TX
centric interrupts, and the handling of jobs.

Say that we wanted to write 200 bytes.  During the driver's write,
we'd check how much space we have in the fifo, then take the min of
that and the job size and start a new TX command with that value as
the size.  The FIFO is 16 4-byte words, so 64 bytes total.  In the
case of a job with 200 bytes, we'd issue a command saying that we have
64 bytes to send, then fill the FIFO with the first 64 bytes.
Finally, we would enable *both* the low-watermark and command-complete
IRQs.

When the TX low-watermark IRQ triggered, the hard IRQ handler would
recognize this as a TX interrupt and wake the writer thread.  The
writer thread would ask the driver "how much space do you have", and
we be told "zero bytes", because there was still a command in
progress.  So the writer thread would enable both interrupts, and wait
again.

At this point in time, the interrupt is effectively stuck.  We are
going to keep spinning until all of the final bytes get sent and the
command becomes complete.  We used to have the low water threshold set
to 4 words, and I discovered by experimentation that the unit will
immediately consume 3 words after a job starts.  So, we might have as
many as 24 bytes pending transmission when the low-water IRQ starts to
fire, meaning that 115.2kbaud, we could end up in this cycle for just
about 2 mSec.

The fix here is to rework how we use these interrupts.  One option
might be to just use the command complete IRQ, and nothing else,
allowing the FIFO to run completely dry every time.  This would work,
but when transmitting large jobs, it would be a bit sub-optimal.

Instead what we do is the following:
1) When write is called, if there is no job in progress, we start a
   new job with the total number of bytes we were asked to transmit,
   and we record this number as "tx_pending_bytes".
2) As we pack bytes into the FIFO, we decrement tx_pending_bytes.
3) When we are done, we have either packed the FIFO completely full, or
   we have packed the entire job into the FIFO and tx_pending_bytes is
   now zero.
4) When we re-enable TX interrupts after the write is done, we only
   enable the low watermark IRQ if there are still bytes to transmit
   in the job.
5) Finally, when reporting how much space we have, we report the
   actual number of free bytes in the fifo if there are still bytes in
   the job to send (tx_pending_bytes > 0).  If we have packed
   everything into the FIFO already, we report 0 if there is still a
   command running, preventing us from attempting to start a new
   command before the old one is complete.  Finally, if there are no
   more bytes to pack, and no command running, we just report the
   space we have in the FIFO.

So we effectively try to start larger jobs, but only use the low-water
IRQ when we still have more bytes for the job to pack into the FIFO.
When we run out, we care only about the command-complete IRQ.

Other smaller changes:

Some of the other smaller changes in this CL:

(1)
Lower the low-watermark to 1 word from 4.  This basically means that
we will wait for the FIFO to hit completely empty before waking up to
pack more bytes.  Experiments show that the micro-controller actually
is holding 3 words when the FIFO hits empty, which means we have about
1.04 mSec before the job becomes complete.  A typical service cycle
for the TX interrupt typically takes < 15uSec to service (including
both handing the IRQ, and waking the TX service thread), so we are
probably fine at this low threshold.

(2)
Move the acknowledgement of the interrupts from the end of the IRQ
handler to the start of the IRQ handler.  The interrupts in this
register all have sticky bits and must be acked; they behave like
edge-triggered interrupts.  In theory, by acking these things at the
end, we set up a race.  If we just packed stuff to transmit, what
happens if that command ends up completing _before_ we manage to ack
the "command complete" interrupt?  The net effect would be that we
lose the interrupt and become stuck.  Practically speaking, this will
_never_ happen, but it is always nice to structure things to be as
safe as possible.

(3)
Mask our interrupt status reads with the interrupt-enabled bits.  The
two interrupt status registers in the unit report the unmasked
interrupt status.  So, if the TX FIFO is low, it will report "low
watermark" in the status even if we do not currently care about the
low water threshold.  So, if we received an RX interrupt when the TX
half was completely idle, we'd still decide that this was a TX
interrupt as well and attempt to signal a TX thread (who was not
actually waiting).

(4)
Ack our low-watermark IRQ after filling the FIFO.  Oddly enough, the
low-watermark IRQ status bit seems to be sticky.  Once it fires, it
needs to be manually cleared.  When the IRQ handler fires, it will
clear the bit, but the FIFO is still low.  It is unclear if this
interrupt is a true level triggered interrupt, or if it is actually
more of an edge-triggered interrupt asserted by the micro-controller
when the controller takes a word out of the FIFO and notices that the
FIFO is low.

Either way, we will eventually make it to Write, and start to put
bytes in the FIFO, which the unit will immediately read and start to
transmit.  If the interrupt is actually edge triggered and was not
asserted before, it is asserted now.  So, after filling the FIFO up
completely, we used to re-enable interrupts and the low watermark IRQ
would immediately fire (even though the FIFO was still full).
Clearing the bit after filling the FIFO on the TX thread prevents us
from taking an additional spurious interrupt.

Note:  If the interrupt _is_ actually edge triggered, we _technically_
just set up race where the FIFO might completely drain and then we end
up clearing the pending bit in the interrupt status register.  See the
lengthy comment left in code which describes how we avoid this race
condition, regardless of how the underlying hardware behaves.

Change-Id: I610a714826a1c7ef8bf6d0ef75d77d01ccfdb6c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1299755
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Will Drewry <drewry@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>