[fx][qemu] Add option to skip SSH key patching of the ZBI
The default behavior of `fx qemu` is to patch the ZBI to add
~/.ssh/fuchsia_authorized_keys into the BOOTFS for SSH access. This is
the only logic that assumes the ramdisk to provide to QEMU is a ZBI. A
new testing workflow is being considered where an actual initramfs
(containing Fuchsia images) would be provided to an actual Linux kernel
for QEMU, which would invalidate the ZBI assumption. This gives a
convenient escape hatch for that case. (The SSH key patching is also
something we don't need to do generally (e.g., for bringup) anyway.)

Change-Id: I071342592dc268ab1fdefe1f88eda364b55577a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1321986
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>