[packet-formats] Rewrite use of `.and()` to avoid rust lint
An upcoming version of Rust has a lint that suggests migrating from
`an_option.and(...)` that contains a function call to
`an_optoin.and_then(|| ...)`, which will only call the closure if
`an_option` is `Some(...)`.

In this circumstance, this change seems to be material since we would no
longer be calling `.write_obj_front` twice. Looking at
`write_obj_front`, it only returns `None` if the argument type is larger
than the remaining buffer. So hypothetically if `ts_val` was `u64`, and
`ts_echo_reply` was `u8` it could mutate the buffer if there was only u8
bytes left in the buffer. Fortunately both are `u32` so it'd return
`None` in both cases.

Bug: 430076135
Change-Id: I9832f102ce3a179f35ab99d91954fca73f903014
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322642
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>