[netcfg] Update condition for UpdateApplied
Change the condition on UpdateApplied to check to see if the new default
network / new marks are truly new, and if they are not, do not set the
field in the UpdateApplied. This results in unnecessary churn for the
client to re-instantiate the call to watch_default even though the
network itself did not change.

Multiply: policy-management-integration-test
Bug: 431384532
Change-Id: I811bd46a20c3ca06dc93aa74da8457ce79cfd8d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322604
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Sam Balana <sbalana@google.com>