[wlan] Deauthenticate when clearing fw conn state.
When brcmfmac encounters connection failure/timeout, reassociation
failure/timeout, or certain internal errors, it should issue a firmware
command that deauthenticates the STA from the AP. Before this change,
brcmfmac issued a C_DISASSOC command, which results in the firmware
sending an 802.11 disassociate frame to the AP as needed. When a
connection/roam attempt fails with NO_NETWORK error, the disassociation
is sometimes insufficient to clear the connection state, resulting in
subsequent attempts frequently becoming persistently unsuccessful. This
CL changes the logic to send a C_SCB_DEAUTHENTICATE_FOR_REASON command,
which results in an 802.11 deauthenticate frame being sent to the AP as
needed. With this change, a physical device that encounters a
NO_NETWORK error has been seen to recover on the next connection attempt
rather than remaining persistently unable to connect. In addition, the
firmware was observed to send a deauthenticate frame to the AP instead
of a disassociate frame.

This change does not affect disconnect logic used in most disconnects
(e.g. user-initiated disconnect, or AP-initiated disconnect), just in
cases where brcmfmac has decided that the connection/roam attempt has
failed in such a way that the firmware connection state needs to be
cleared.

Bug: 406330083, 369703077
Test: fx test brcmfmac_test brcmfmac_sim_test;
      Antlion tests: beacon_loss_test, vape_interop_test,
      wlan_phy_compliance_11n_2g_ht40lower_open_test
Change-Id: I790374827f49845b6560e24065c53ecc768ba425
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322426
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Karl Ward <karlward@google.com>