[starnix] Add lock-level to DeviceRegistry
This change introduces lock-level checks to the `DeviceRegistry` by
making its internal state an `OrderedMutex`. This ensures that the
registry is always accessed with the correct locks held, preventing
potential deadlocks and race conditions.

To support this, the `DeviceRegistry`s methods now require a `Locked`
context to be passed in. This allows the `OrderedMutex` to verify
that the caller is holding the necessary locks before accessing the
registrys state.

The call sites of the `DeviceRegistry`s methods have been updated to
pass the `Locked` context.

Change-Id: Ia8876b5908b5f0cef64c7c16d84502081098e6c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1316044
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>