[starnix] Add lock-level to FutexTable
This change introduces lock-level checks to the `FutexTable` by
making its internal state an `OrderedMutex`. This ensures that the
futex table is always accessed with the correct locks held, preventing
potential deadlocks and race conditions.

To support this, the `FutexTable`s methods now require a `Locked`
context to be passed in. This allows the `OrderedMutex` to verify
that the caller is holding the necessary locks before accessing the
futex tables state.

The call sites of the `FutexTable`s methods have been updated to
pass the `Locked` context.

Change-Id: I15bd6e9e91ce52a313942b4421aad6a23c10441e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317005
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>