[starnix] Add unregister hook to DeviceOps
This change introduces an 'unregister' operation on the 'DeviceOps'
trait. This allows devices to perform necessary cleanup when they are
unregistered from the system.

A key challenge is that the 'unregister' method must be called with a
specific lock context to prevent potential deadlocks. To ensure this
is always the case, this change utilizes a delayed release mechanism.

When a device is unregistered, the call to its 'unregister' method is
deferred until the system can guarantee the correct lock state,
ensuring a safe and orderly cleanup process.

Change-Id: Iac2e65c7fba380f6b70a6dddc8ca3be30d3bad2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1319244
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>