[starnix] Refactor FileObject to prepare for owned close
This change refactors the `FileObject` by splitting its fields into a
new `FileObjectState` struct. The `FileObject` now contains the `ops`
and the `FileObjectState`.

This is a preparatory refactoring to enable a better destruction
pattern for `FileOps`. The signature of the `FileOps::close` method
will be changed to take ownership of the ops object (e.g., `Box<Self>`).

Passing an owned `self` to `close` is not possible if the method is
also passed a reference to the containing `FileObject`, as this would
involve moving out of a borrowed context.

By splitting the state, the `close` method can be called with the
`FileObjectState` without borrowing the `FileObject` that contains the
`ops`. This paves the way for changing the `close` signature to take
ownership.

Change-Id: Ide187b2804b0798bec694491414bd1ec904ccb65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1319264
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>