[starnix] Change FileOps::close to take self by value
This change modifies the signature of `FileOps::close` to take `self`
by value (as `Box<Self>`) instead of by reference. This allows the
`close` method to consume the `FileOps` object, enabling a cleaner
destruction pattern.

All implementations of `FileOps::close` have been updated to reflect
this new signature.

Additionally, a new marker trait, `CloseFreeSafe`, is introduced.
This trait marks `FileOps` implementations that can be safely wrapped
in containers like `Arc`. When a wrapper implements `FileOps`, its
`close` method receives an owned `Box` of the wrapper. It is not
possible to delegate this call to the wrapped object, as this would
require taking ownership of the wrapped object, which the wrapper
cannot provide.

The `CloseFreeSafe` trait signifies that the wrapped object does not
perform any essential cleanup in its `close` method, and thus it is
safe not to call it in this wrapper scenario.

Change-Id: Iaf141e6554c53fa5f2d5b9279a98dfc6a0d069ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1319304
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>