[starnix] Change DeviceOps::open to take a NamespaceNode
This change modifies the signature of `DeviceOps::open` to take a
`&NamespaceNode` instead of a `&FsNode`.

This is a preparatory refactoring to enable a correct implementation
of `ptmx`. The `ptmx` device must act as a symbolic link to
`/dev/pts/ptmx` when it is not on a `devpts` filesystem.

To implement this behavior, the `open` logic for `ptmx` needs to
inspect the filesystem hierarchy it is being opened from. The `NamespaceNode`
provides this context.

The `FsNode::open` method signature was also updated to pass the
`NamespaceNode` down to the underlying device or filesystem
implementation.

Change-Id: I0e7884766a170aa50c59bfb112228e9d9acf40e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1320084
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>