[fxfs] Hold epoch refs for entire page in.
Previously we were adding the epoch ref long after we'd verified that
the request should be done due to the scope being open and the lock for
the pager's reference to the file had been released.

Now the epoch ref is taken early on in the request lifetimes. Moves the
ref and the active guard are inside the `PagerRange` struct to make it
easy to track that they live as long as the request does.

Bug: 430394762
Change-Id: I7493deb151112198d0b7bc490b4a6e173a19e69a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1321824
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>