[machina] Add InteractiveDebianGuest FIDL and Impl
This is both a quality-of-life wrapper around the GuestManager, as well
as a necessary wrapper around the associated gRPC client comms that
occur over vsock to the guest. The primary consumer will be Starnix,
with Netstack as a potential secondary consumer in the future.

Bug: 418575059
Change-Id: Icced63b0e414d7ef3bb185f47f249755b2d7bbfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281784
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Aaron Knobloch <aknobloch@google.com>
Fuchsia-Auto-Submit: Aaron Knobloch <aknobloch@google.com>