[ffx] Fix deadlock with concurrent ssh connections
Change how connections are managed by NetworkConnector: Instead
of a Mutex lock, we use ArcSwap, which is a lock-free mechanism
for updating a field with an Arc.

To prevent multiple tasks from trying to establish a connection
at the same time, we introduce a semaphore, which only protects
the code _establishing_ the connection (as opposed to the previous
Mutex, which controlled all access to the Connection).

We also change the semantics of NetworkConnector::connection() slightly:
instead of returning an Result<Option<Connection>>, we simply use an
Err result to indicate a failure to establish a connection, returning
Result<Connection>. (Well, an Arc<Connection>.)

Change-Id: I6a6a696432157fe9133c7a25485df36da84ee657
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1322632
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>