[trace] Changing duration type
This changes the fidl TraceOptions to store the
duration as a zx.Duration, which represents nanoseconds.

This also changes the command line for `ffx trace start` to
make duration a u32 for seconds.

This is done as clean up and improving consistency of the
API.

Change-Id: Iac20d3b84600e90be09f00affedf468a27b45757
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1323672
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
API-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>