[bt][a2dp] Move encoding tests to inband_source
Do not error and stop the whole profile when we cannot encode
SBC if we are not going to be an audio source.

Adjust availability of CodecFactory to optional as we may not need to
do encoding locally if i.e. we are using offload audio encoding.

Bug: 278283913
Test: moved & updated tests, build without CodecFactory available
Change-Id: I96b26d8f73246dbd158a521b8e346cd2028d7a80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021472
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>