[performance][ffx] Removing TraceTask.onComplete callback
This removes the callback for on_complete in the TraceTask. It was
used for tasks that are run via the ffx daemon so the tasks would be
dropped.

This is done anyway, so the callback was redundant.

Removing this makes it easier to enable using TraceTask on the device
to support disconnected trace collection.

Bug: 428951658
Change-Id: Ia5f2b63e5c1ad0a40e78bf26670efe8dc3037636
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324724
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>