[node-removal] don't add stopped nodes
discovered a bug while touching component shutdown code.
if a node is already in a stopped state, it's registered.
however, it's Stopped so Notify is never called and it's
never removed from the tracking count which causes
OnRemovalTimeout to hang. This CL avoids that behavior
by not registering a node if its state is Stopped.

Change-Id: I9926edc016eebe72a8629fec0f05b46e8604120b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324584
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Olivia Nordquist <nolivia@google.com>