[display] Inline testing/client-utils into bin/display-test.
client-utils is not currently suitable for display stack consumers. The
library has not been maintaining and lacks the level of testing we
expect in code intended for broad consumption.

This CL reflects the fact that display-test is the only consumer of that
code.

Bug: 323437941
Test: fx build (no behavior changes)
Change-Id: Iaa19e68e0b8f9b7e166a48c59622a930db2a8a22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1324489
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>