[assembly] Rework how //build/images/updates collects package manifests from bundles
This factors out to separate files the specifics of how the lists of
package manifests for each package set, and the main product bundle,
are gathered depending on if it's GN or Bazel-based.

The logic within //build/images/updates no longer various based on
whether the assembly is implemented in GN or in Bazel.

The lists of the names of the packages in various package sets are now
actually lists of names, as opposed to paths which usually, but not
always, ended in the package name.  'fx test' and the utility shell
function 'update-if-in-base' have been updated to properly use the
lists of names.

Further, several now-unused files were removed, and all vars in
//build/images/paths.gni that were only used in //build/images/updates
were removed from the gni file.

Change-Id: I1138fb0faa90e269051ec295446627e50247218c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1319065
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>